+++ ParseWiki/ParseWiki/Sources/IWithId.cs
@@ -2,6 +2,6 @@ namespace ParseWiki.Sources
 {
     public interface IWithId
     {
-        public int Id { get; }
+        public long Id { get; }
     }
 }
